common: fix trivial warnings in src and internal folders
replacing: asin->asinf, acos->acosf, atan->atanf, atan2->atan2f,
ceilf->ceilf, exp->expf fabs->fabsf, floor->floorf, log->logf, pow->powf,
sin->sinf, cos->cosf, coshf->coshf, tan->tanf, sqrt->sqrtf

when the result is needed as a float instead of double

+++ dev/src/common/core/algorithms.h
@@ -89,7 +89,7 @@ RED_INLINE T Abs(const T& a)
 template <>
 RED_INLINE Float Abs< Float >(const Float& a)				
 { 
-	return fabs( a ) ;
+	return fabsf( a ) ;
 }
